requirements.txt
imageio
matplotlib
numpy~=1.26.2
pandas~=2.1.4
Pillow~=10.0.1
scikit_image
scipy
timm
torch
torchsummary
torchvision~=0.16.2
tqdm
wget
thop~=0.1.1-2209072238
einops~=0.8.0